Add docs `Date::class`.
